+++ packages/mobile/components/Themed.tsx
@@ -5,7 +5,7 @@
 
 import { Text as DefaultText, View as DefaultView, useColorScheme } from 'react-native';
 
-import Colors from '../constants/colors';
+import { colors } from '../constants/color';
 
 type ThemeProps = {
   lightColor?: string;
@@ -17,7 +17,7 @@ export type ViewProps = ThemeProps & DefaultView['props'];
 
 export function useThemeColor(
   props: { light?: string; dark?: string },
-  colorName: keyof typeof Colors.light & keyof typeof Colors.dark,
+  colorName: keyof typeof colors.light & keyof typeof colors.dark,
 ) {
   const theme = useColorScheme() ?? 'light';
   const colorFromProps = props[theme];
@@ -25,7 +25,7 @@ export function useThemeColor(
   if (colorFromProps) {
     return colorFromProps;
   } else {
-    return Colors[theme][colorName];
+    return colors[theme][colorName];
   }
 }
