+++ packages/mobile/components/edit-screen.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet } from 'react-native';
 
-import Colors from '../constants/colors';
+import { colors } from '../constants/color';
 import { ExternalLink } from './ExternalLink';
 import { MonoText } from './StyledText';
 import { Text, View } from './Themed';
@@ -32,7 +32,7 @@ export default function Leaderboard({ path }: { path: string }) {
           style={styles.helpLink}
           href="https://docs.expo.io/get-started/create-a-new-app/#opening-the-app-on-your-phonetablet"
         >
-          <Text style={styles.helpLinkText} lightColor={Colors.light.tint}>
+          <Text style={styles.helpLinkText} lightColor={colors.light.tint}>
             and the best potter
           </Text>
         </ExternalLink>
